+++ tests/FSharp.CosmosDb.Analyzer.Tests/InsertTests.fs
@@ -1,31 +1,30 @@
-module Tests
+module InsertTests
 
 open Expecto
-open System
 open FSharp.CosmosDb.Analyzer
 
 [<Tests>]
 let tests =
-    testList "Connection object discovery"
+    testList "Insert API can be analyzed"
         [ test "Finds all the operations in a file" {
-              match context (find "../samples/simpleSample.fs") with
+              match context (find "../samples/insertSample.fs") with
               | None -> failwith "Could not load test script"
               | Some context ->
                   let ops = CosmosCodeAnalysis.findOperations context
-                  Expect.equal (List.length ops) 1 "Found one operation block"
+                  Expect.equal 1 (List.length ops) "Found one operation block"
           }
 
-          test "Found operation should have 2 analysable bits" {
-              match context (find "../samples/simpleSample.fs") with
+          test "Found operation should have correct number of analysable bits" {
+              match context (find "../samples/insertSample.fs") with
               | None -> failwith "Could not load test script"
               | Some context ->
                   let ops = CosmosCodeAnalysis.findOperations context
                   let head = List.exactlyOne ops
-                  Expect.equal (List.length head.blocks) 2 "Found two things to analyse"
+                  Expect.equal 2 (List.length head.blocks) "Found two things to analyse"
           }
 
           test "DatabaseId should match expected" {
-              match context (find "../samples/simpleSample.fs") with
+              match context (find "../samples/insertSample.fs") with
               | None -> failwith "Could not load test script"
               | Some context ->
                   let ops = CosmosCodeAnalysis.findOperations context
@@ -40,11 +39,11 @@ let tests =
                           | CosmosAnalyzerBlock.DatabaseId(dbId, _) -> dbId
                           | _ -> failwith "Should've found the DatabaseId operation")
 
-                  Expect.equal dbId.Value "UserDb" "DatabaseId matches the one in code"
+                  Expect.equal "UserDb" dbId.Value "DatabaseId matches the one in code"
           }
 
           test "ContainerName should match expected" {
-              match context (find "../samples/simpleSample.fs") with
+              match context (find "../samples/insertSample.fs") with
               | None -> failwith "Could not load test script"
               | Some context ->
                   let ops = CosmosCodeAnalysis.findOperations context
@@ -59,5 +58,5 @@ let tests =
                           | CosmosAnalyzerBlock.ContainerName(containerName, _) -> containerName
                           | _ -> failwith "Should've found the ContainerName operation")
 
-                  Expect.equal container.Value "UserContainer" "ContainerName matches the one in code"
+                  Expect.equal "UserContainer" container.Value "ContainerName matches the one in code"
           } ]