+++ tests/samples/insertSample.fs
@@ -1,13 +1,22 @@
-module FirstSample
+module InsertSamples
 
 open FSharp.CosmosDb
 
 let key = "testing"
 let endpointUrl = "https://localhost"
 
-let findUsers() =
+[<CLIMutable>]
+type User =
+    { Id: string
+      Name: string }
+
+let insertUser() =
     endpointUrl
     |> Cosmos.host
     |> Cosmos.connect key
     |> Cosmos.database "UserDb"
     |> Cosmos.container "UserContainer"
+    |> Cosmos.insert
+        { Id = ""
+          Name = "" }
+    |> Cosmos.execAsync